dash_tools -> utils/dash_utils (plus other renames)
