+++ examples/video_mode/example_video_mode_opx.py
@@ -1,6 +1,16 @@
 # %% Imports
-from qm import QuantumMachinesManager
-import logging
+from qua_dashboards.video_mode.utils.dash_utils import *
+import numpy as np
+from matplotlib import pyplot as plt
+from qua_dashboards.video_mode.voltage_parameter import *
+from qua_dashboards.video_mode.sweep_axis import *
+from qua_dashboards.video_mode.data_acquirers import *
+from qua_dashboards.video_mode import scan_modes
+from qua_dashboards.video_mode.inner_loop_actions.basic_inner_loop_action import (
+    BasicInnerLoopAction,
+)
+# from qua_dashboards.video_mode import VideoModeApp
+
 from quam.components import (
     BasicQuam,
     SingleChannel,
