+++ examples/video_mode/example_video_mode_random.py
