+++ src/qua_dashboards/video_mode/__init__.py
@@ -1,5 +1,5 @@
 from .video_mode_component import *
-from .dash_tools import *
+from .utils.dash_utils import *
 from .sweep_axis import SweepAxis
 from .voltage_parameter import *
 from .inner_loop_actions import *
@@ -17,7 +17,7 @@ from .data_registry import (
 
 __all__ = [
     *video_mode_component.__all__,
-    *dash_tools.__all__,
+    *dash_utils.__all__,
     "SweepAxis",
     *voltage_parameter.__all__,
     *inner_loop_actions.__all__,
