+++ src/qua_dashboards/video_mode/inner_loop_actions/inner_loop_action.py
@@ -1,3 +1,4 @@
+from dash import Dash
 from qua_dashboards.core import BaseUpdatableComponent
 from qua_dashboards.utils.qua_types import QuaVariableFloat
 
@@ -14,6 +15,12 @@ class InnerLoopAction(BaseUpdatableComponent, ABC):
     def __call__(
         self, x: QuaVariableFloat, y: QuaVariableFloat
     ) -> Tuple[QuaVariableFloat, QuaVariableFloat]:
+        pass
+
+    def get_layout(self):
+        return None
+
+    def register_callbacks(self, app: Dash):
         pass
 
     def initial_action(self):
