+++ src/qua_dashboards/video_mode/shared_viewer_component.py
@@ -6,7 +6,7 @@ import xarray as xr
 from dash import Dash, Input, Output, State, dcc, html
 
 from qua_dashboards.core import BaseComponent
-from qua_dashboards.video_mode.dash_tools import xarray_to_plotly
+from qua_dashboards.video_mode.utils.dash_utils import xarray_to_plotly
 from qua_dashboards.video_mode import data_registry
 
 from qua_dashboards.video_mode.utils.annotation_utils import (
